[FIX] Removing member initializer lists in seqcons2.
These do not work with VS11.

+++ extras/apps/seqcons2/seqcons_app.cpp
@@ -279,7 +279,7 @@ void SeqConsAppImpl::performConsensusAlignment(bool useContigID, bool usePositio
     if (options.verbosity >= 3)
         caOptions.verbosity = 3;
 
-    caOptions.overlapMaxErrorRate = options.overlapMaxErrorRate;
+    caOptions.overlapMaxErrorRate = (int)options.overlapMaxErrorRate;
     caOptions.overlapMinLength = options.overlapMinLength;
     caOptions.overlapMinCount = options.overlapMinCount;
     caOptions.posDelta = options.overlapWindowSize;
